// examples/app-crm/src/graphql/schema.types.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  DateTime: { input: any; output: any };
};
// ...
export type BusinessType = "B2B" | "B2C" | "B2G";

export type CheckListItem = {
  checked: Scalars["Boolean"]["output"];
  title: Scalars["String"]["output"];
};

export type ChecklistItemInput = {
  checked: Scalars["Boolean"]["input"];
  title: Scalars["String"]["input"];
};

export type Company = {
  avatarUrl?: Maybe<Scalars["String"]["output"]>;
  businessType?: Maybe<BusinessType>;
  companySize?: Maybe<CompanySize>;
  contacts: CompanyContactsConnection;
  contactsAggregate: Array<CompanyContactsAggregateResponse>;
  country?: Maybe<Scalars["String"]["output"]>;
  createdAt: Scalars["DateTime"]["output"];
  createdBy: User;
  deals: CompanyDealsConnection;
  dealsAggregate: Array<CompanyDealsAggregateResponse>;
  id: Scalars["ID"]["output"];
  industry?: Maybe<Industry>;
  name: Scalars["String"]["output"];
  notes: CompanyNotesConnection;
  notesAggregate: Array<CompanyNotesAggregateResponse>;
  salesOwner: User;
  totalRevenue?: Maybe<Scalars["Int"]["output"]>;
  updatedAt: Scalars["DateTime"]["output"];
  updatedBy?: Maybe<User>;
  website?: Maybe<Scalars["String"]["output"]>;
};
// ...
export type CompanyContactsAggregateResponse = {
  avg?: Maybe<CompanyContactsAvgAggregate>;
  count?: Maybe<CompanyContactsCountAggregate>;
  groupBy?: Maybe<CompanyContactsAggregateGroupBy>;
  max?: Maybe<CompanyContactsMaxAggregate>;
  min?: Maybe<CompanyContactsMinAggregate>;
  sum?: Maybe<CompanyContactsSumAggregate>;
};

export type CompanyContactsAvgAggregate = {
  id?: Maybe<Scalars["Float"]["output"]>;
  score?: Maybe<Scalars["Float"]["output"]>;
};

export type CompanyContactsConnection = {
  /** Array of nodes. */
  nodes: Array<Contact>;
  /** Paging information */
  pageInfo: OffsetPageInfo;
  /** Fetch total count of records */
  totalCount: Scalars["Int"]["output"];
};
// ...
export type CompanyDealsAggregateResponse = {
  avg?: Maybe<CompanyDealsAvgAggregate>;
  count?: Maybe<CompanyDealsCountAggregate>;
  groupBy?: Maybe<CompanyDealsAggregateGroupBy>;
  max?: Maybe<CompanyDealsMaxAggregate>;
  min?: Maybe<CompanyDealsMinAggregate>;
  sum?: Maybe<CompanyDealsSumAggregate>;
};
// ...
export type CompanyDealsConnection = {
  /** Array of nodes. */
  nodes: Array<Deal>;
  /** Paging information */
  pageInfo: OffsetPageInfo;
  /** Fetch total count of records */
  totalCount: Scalars["Int"]["output"];
};
// ...
export type CompanyNotesAggregateResponse = {
  avg?: Maybe<CompanyNotesAvgAggregate>;
  count?: Maybe<CompanyNotesCountAggregate>;
  groupBy?: Maybe<CompanyNotesAggregateGroupBy>;
  max?: Maybe<CompanyNotesMaxAggregate>;
  min?: Maybe<CompanyNotesMinAggregate>;
  sum?: Maybe<CompanyNotesSumAggregate>;
};

export type CompanyNotesAvgAggregate = {
  id?: Maybe<Scalars["Float"]["output"]>;
};

export type CompanyNotesConnection = {
  /** Array of nodes. */
  nodes: Array<CompanyNote>;
  /** Paging information */
  pageInfo: OffsetPageInfo;
  /** Fetch total count of records */
  totalCount: Scalars["Int"]["output"];
};
// ...
export type CompanySize = "ENTERPRISE" | "LARGE" | "MEDIUM" | "SMALL";
// ...
export type Contact = {
  avatarUrl?: Maybe<Scalars["String"]["output"]>;
  company: Company;
  createdAt: Scalars["DateTime"]["output"];
  createdBy: User;
  deals: ContactDealsConnection;
  email: Scalars["String"]["output"];
  id: Scalars["ID"]["output"];
  jobTitle?: Maybe<Scalars["String"]["output"]>;
  name: Scalars["String"]["output"];
  notes: ContactNotesConnection;
  phone?: Maybe<Scalars["String"]["output"]>;
  salesOwner: User;
  score?: Maybe<Scalars["Int"]["output"]>;
  stage: ContactStage;
  status: ContactStatus;
  timezone?: Maybe<Scalars["String"]["output"]>;
  updatedAt: Scalars["DateTime"]["output"];
  updatedBy?: Maybe<User>;
};
// ...
export type ContactDealsConnection = {
  /** Array of nodes. */
  nodes: Array<Deal>;
  /** Paging information */
  pageInfo: OffsetPageInfo;
  /** Fetch total count of records */
  totalCount: Scalars["Int"]["output"];
};
// ...
export type ContactNote = {
  contact: Contact;
  createdAt: Scalars["DateTime"]["output"];
  createdBy: User;
  id: Scalars["ID"]["output"];
  note: Scalars["String"]["output"];
  updatedAt: Scalars["DateTime"]["output"];
  updatedBy?: Maybe<User>;
};
// ...
export type ContactNotesConnection = {
  /** Array of nodes. */
  nodes: Array<ContactNote>;
  /** Paging information */
  pageInfo: OffsetPageInfo;
  /** Fetch total count of records */
  totalCount: Scalars["Int"]["output"];
};

export type ContactSort = {
  direction: SortDirection;
  field: ContactSortFields;
  nulls?: InputMaybe<SortNulls>;
};

export type ContactSortFields =
  | "createdAt"
  | "email"
  | "id"
  | "jobTitle"
  | "name"
  | "phone"
  | "score"
  | "stage"
  | "status"
  | "timezone"
  | "updatedAt";

/** Contact stage */
export type ContactStage = "CUSTOMER" | "LEAD" | "SALES_QUALIFIED_LEAD";
// ...
export type ContactStatus =
  | "CHURNED"
  | "CONTACTED"
  | "INTERESTED"
  | "LOST"
  | "NEGOTIATION"
  | "NEW"
  | "QUALIFIED"
  | "UNQUALIFIED"
  | "WON";
// ...
export type Deal = {
  closeDateDay?: Maybe<Scalars["Int"]["output"]>;
  closeDateMonth?: Maybe<Scalars["Int"]["output"]>;
  closeDateYear?: Maybe<Scalars["Int"]["output"]>;
  company: Company;
  companyId: Scalars["ID"]["output"];
  createdAt: Scalars["DateTime"]["output"];
  createdBy: User;
  dealContact: Contact;
  dealOwner: User;
  dealOwnerId: Scalars["ID"]["output"];
  id: Scalars["ID"]["output"];
  notes: Scalars["String"]["output"];
  stage?: Maybe<DealStage>;
  stageId?: Maybe<Scalars["ID"]["output"]>;
  title: Scalars["String"]["output"];
  updatedAt: Scalars["DateTime"]["output"];
  updatedBy?: Maybe<User>;
  value?: Maybe<Scalars["Float"]["output"]>;
};
// ...
export type DealStage = {
  createdAt: Scalars["DateTime"]["output"];
  createdBy: User;
  deals: Array<Deal>;
  dealsAggregate: Array<DealStageDealsAggregateResponse>;
  id: Scalars["ID"]["output"];
  title: Scalars["String"]["output"];
  updatedAt: Scalars["DateTime"]["output"];
  updatedBy?: Maybe<User>;
};
// ...
export type DealStageDealsAggregateResponse = {
  avg?: Maybe<DealStageDealsAvgAggregate>;
  count?: Maybe<DealStageDealsCountAggregate>;
  groupBy?: Maybe<DealStageDealsAggregateGroupBy>;
  max?: Maybe<DealStageDealsMaxAggregate>;
  min?: Maybe<DealStageDealsMinAggregate>;
  sum?: Maybe<DealStageDealsSumAggregate>;
};
// ...
export type Event = {
  category: EventCategory;
  color: Scalars["String"]["output"];
  createdAt: Scalars["DateTime"]["output"];
  createdBy: User;
  description: Scalars["String"]["output"];
  endDate: Scalars["DateTime"]["output"];
  id: Scalars["ID"]["output"];
  participants: Array<User>;
  startDate: Scalars["DateTime"]["output"];
  title: Scalars["String"]["output"];
  updatedAt: Scalars["DateTime"]["output"];
  updatedBy?: Maybe<User>;
};
// ...
export type Industry =
  | "AEROSPACE"
  | "AGRICULTURE"
  | "AUTOMOTIVE"
  | "CHEMICALS"
  | "CONSTRUCTION"
  | "DEFENSE"
  | "EDUCATION"
  | "ENERGY"
  | "FINANCIAL_SERVICES"
  | "FOOD_AND_BEVERAGE"
  | "GOVERNMENT"
  | "HEALTHCARE"
  | "HOSPITALITY"
  | "INDUSTRIAL_MANUFACTURING"
  | "INSURANCE"
  | "LIFE_SCIENCES"
  | "LOGISTICS"
  | "MEDIA"
  | "MINING"
  | "NONPROFIT"
  | "OTHER"
  | "PHARMACEUTICALS"
  | "PROFESSIONAL_SERVICES"
  | "REAL_ESTATE"
  | "RETAIL"
  | "TECHNOLOGY"
  | "TELECOMMUNICATIONS"
  | "TRANSPORTATION"
  | "UTILITIES";
// ...
export type OffsetPageInfo = {
  /** true if paging forward and there are more records. */
  hasNextPage?: Maybe<Scalars["Boolean"]["output"]>;
  /** true if paging backwards and there are more records. */
  hasPreviousPage?: Maybe<Scalars["Boolean"]["output"]>;
};
// ...
export type Role = "ADMIN" | "SALES_INTERN" | "SALES_MANAGER" | "SALES_PERSON";
// ...
export type Task = {
  checklist: Array<CheckListItem>;
  comments: TaskCommentsConnection;
  commentsAggregate: Array<TaskCommentsAggregateResponse>;
  completed: Scalars["Boolean"]["output"];
  createdAt: Scalars["DateTime"]["output"];
  createdBy: User;
  description?: Maybe<Scalars["String"]["output"]>;
  dueDate?: Maybe<Scalars["DateTime"]["output"]>;
  id: Scalars["ID"]["output"];
  stage?: Maybe<TaskStage>;
  stageId?: Maybe<Scalars["ID"]["output"]>;
  title: Scalars["String"]["output"];
  updatedAt: Scalars["DateTime"]["output"];
  updatedBy?: Maybe<User>;
  users: Array<User>;
  usersAggregate: Array<TaskUsersAggregateResponse>;
};
// ...
export type User = {
  avatarUrl?: Maybe<Scalars["String"]["output"]>;
  companies: UserCompaniesConnection;
  contacts: UserContactsConnection;
  createdAt: Scalars["DateTime"]["output"];
  createdBy?: Maybe<User>;
  deals: UserDealsConnection;
  email: Scalars["String"]["output"];
  events: UserEventsConnection;
  id: Scalars["ID"]["output"];
  jobTitle?: Maybe<Scalars["String"]["output"]>;
  name: Scalars["String"]["output"];
  phone?: Maybe<Scalars["String"]["output"]>;
  role: Role;
  tasks: UserTasksConnection;
  timezone?: Maybe<Scalars["String"]["output"]>;
  updatedAt: Scalars["DateTime"]["output"];
  updatedBy?: Maybe<User>;
};
// ...
export type UserCompaniesConnection = {
  /** Array of nodes. */
  nodes: Array<Company>;
  /** Paging information */
  pageInfo: OffsetPageInfo;
  /** Fetch total count of records */
  totalCount: Scalars["Int"]["output"];
};

export type UserConnection = {
  /** Array of nodes. */
  nodes: Array<User>;
  /** Paging information */
  pageInfo: OffsetPageInfo;
  /** Fetch total count of records */
  totalCount: Scalars["Int"]["output"];
};

export type UserContactsConnection = {
  /** Array of nodes. */
  nodes: Array<Contact>;
  /** Paging information */
  pageInfo: OffsetPageInfo;
  /** Fetch total count of records */
  totalCount: Scalars["Int"]["output"];
};

export type UserCreateInput = {
  email: Scalars["String"]["input"];
  jobTitle: Scalars["String"]["input"];
  name: Scalars["String"]["input"];
  phone: Scalars["String"]["input"];
  role: Role;
  timezone: Scalars["String"]["input"];
};

export type UserDealsConnection = {
  /** Array of nodes. */
  nodes: Array<Deal>;
  /** Paging information */
  pageInfo: OffsetPageInfo;
  /** Fetch total count of records */
  totalCount: Scalars["Int"]["output"];
};
// ...
export type UserEventsConnection = {
  /** Array of nodes. */
  nodes: Array<Event>;
  /** Paging information */
  pageInfo: OffsetPageInfo;
  /** Fetch total count of records */
  totalCount: Scalars["Int"]["output"];
};
// ...
export type UserTasksConnection = {
  /** Array of nodes. */
  nodes: Array<Task>;
  /** Paging information */
  pageInfo: OffsetPageInfo;
  /** Fetch total count of records */
  totalCount: Scalars["Int"]["output"];
};